# flowo-backend/docs/swagger.yaml
basePath: /
definitions:
  dto.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      user_id:
        type: integer
    required:
    - product_id
    - quantity
    - user_id
    type: object
  dto.CartItemResponse:
    properties:
      description:
        type: string
      effective_price:
        description: ImageURL      string  `json:"image_url,omitempty"`
        type: number
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: number
    type: object
  dto.CreatePricingRuleRequest:
    properties:
      adjustment_type:
        enum:
        - percentage_discount
        - fixed_discount
        - override_price
        type: string
      adjustment_value:
        type: number
      applicable_flower_type_id:
        type: integer
      applicable_product_id:
        type: integer
      applicable_product_status:
        type: string
      is_active:
        type: boolean
      priority:
        type: integer
      rule_name:
        type: string
      special_day_id:
        type: integer
      time_of_day_end:
        type: string
      time_of_day_start:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - adjustment_type
    - adjustment_value
    - priority
    - rule_name
    type: object
...
  dto.RemoveCartItemRequest:
    properties:
      product_id:
        type: integer
      user_id:
        type: integer
    required:
    - product_id
    - user_id
    type: object
...
  dto.UpdateCartItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      user_id:
        type: integer
    required:
    - product_id
    - quantity
    - user_id
    type: object
...
  model.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@example.com
    name: API Support Team
    url: http://www.example.com/support
  description: A modern RESTful API for managing your todos efficiently. This API
    provides comprehensive endpoints for creating, reading, updating, and deleting
    todo items.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Todo List API
  version: "1.0"
paths:
  /api/v1/cart/{userID}:
    get:
      description: Retrieve all items in the cart for a given user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CartItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get cart items for user
      tags:
      - cart
  /api/v1/cart/add:
    post:
      consumes:
      - application/json
      description: Add a product with quantity to the user's cart. Stock will be updated
        accordingly.
      parameters:
      - description: Add to cart request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Add product to cart
      tags:
      - cart
  /api/v1/cart/remove:
    delete:
      consumes:
      - application/json
      description: Remove a product from the user's cart
      parameters:
      - description: Remove cart item request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Remove product from cart
      tags:
      - cart
  /api/v1/cart/update:
    put:
      consumes:
      - application/json
      description: Update the quantity of an existing cart item
      parameters:
      - description: Update cart item request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update quantity of a product in cart
      tags:
      - cart
...
  /api/v1/pricing/rule:
    post:
      consumes:
      - application/json
      description: Admin adds a new dynamic pricing rule
      parameters:
      - description: New Pricing Rule
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePricingRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Add new pricing rule
      tags:
      - pricing
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...